farmer_backend/requirements.txt
aiohttp
functions-framework
python-dotenv
openai
celery
fastapi
redis
magika
pillow
opencv-python
qdrant-client
python-pptx
python-docx
google-generativeai
psycopg2-binary
pandas
uvicorn 
fastapi
matplotlib
websockets
soundfile
websocket
python-multipart